restructure directory structure
