+++ .travis.yml
@@ -9,7 +9,7 @@ services:
  - postgresql
 before_script:
   - psql -c "CREATE DATABASE shoppinglist;" -U postgres
-  - psql -c "CREATE USER shoppinglist WITH PASSWORD 'bar';" -U postgres
+  - psql -c "CREATE USER shoppinglist WITH PASSWORD 'Andela100';" -U postgres
   - alembic upgrade head
 script:
   - python3.6 tests.py
